pyproject.toml
[project]
name = "cuai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anomalib>=2.0.0",
    "kornia>=0.8.1",
    "lightning>=2.5.2",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "onnxruntime>=1.22.1",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.1",
    "pywebview>=5.4",
    "ruff>=0.12.1",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.0",
    "timm>=1.0.17",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
]
